test-requirements.txt
pytest>=2.7.0, <4.0.0
pytest-runner>=2.0.0, <3.0.0

httmock>=1.2.0, <2.0.0
PyHamcrest>=1.8.0, <1.10.0
